updating repositories

+++ Supplier.Domain/Services/SupplierService.cs
@@ -6,6 +6,7 @@ using SupplierProject.Domain.Models;
 using SupplierProject.Domain.Validations;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -15,11 +16,16 @@ namespace SupplierProject.Domain.Services
     {
         private readonly IMapper _mapper;
         private readonly ISupplierRepository _supplierRepository;
+        private readonly IAddressRepository _addressRepository;
 
-        public SupplierService(ISupplierRepository supplierRepository, IMapper mapper)
+        public SupplierService(
+            ISupplierRepository supplierRepository,
+            IMapper mapper, 
+            IAddressRepository addressRepository)
         {
             _supplierRepository = supplierRepository;
             _mapper = mapper;
+            _addressRepository = addressRepository;
         }
 
         public async Task<IEnumerable<SupplierDTO>> GetAll()
@@ -56,6 +62,17 @@ namespace SupplierProject.Domain.Services
 
         public async Task<bool> Destroy(Guid id)
         {
+            var hasProduct = _supplierRepository.GetSupplierAndAddressAndProducts(id).Result.Products.Any();
+
+            if (hasProduct) return false;
+
+            var address = await _addressRepository.GetAddressBySupplierId(id);
+
+            if (address != null)
+            {
+                await _addressRepository.Destroy(address.Id);
+            }
+
             var result = await _supplierRepository.Destroy(id);
 
             if (result == 0) return false;
@@ -71,6 +88,12 @@ namespace SupplierProject.Domain.Services
         public async Task<SupplierDTO> GetSupplierAndAddressAndProducts(Guid id)
         {
             return _mapper.Map<SupplierDTO>(await _supplierRepository.GetSupplierAndAddressAndProducts(id));
+        }
+
+        public void Dispose()
+        {
+            _supplierRepository?.Dispose();
+            _addressRepository?.Dispose();
         }
     }
 }
